refactor(query): define _internal methods
