+++ projects/client/src/lib/features/query/QueryClientProvider.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { type QueryClient } from "@tanstack/svelte-query";
   import { PersistQueryClientProvider } from "@tanstack/svelte-query-persist-client";
-  import { createPersister } from "./createPersister";
+  import { createPersister } from "./_internal/createPersister";
 
   const { children, client }: ChildrenProps & { client: QueryClient } =
     $props();
